requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-cleanup==8.0.0
django-js-asset==2.1.0
django-mdeditor==0.1.20
django-resized==1.0.2
idna==3.4
Markdown==3.4.4
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.1.1
urllib3==2.0.4
webencodings==0.5.1
